+++ src/config/config_definition.cc
@@ -534,6 +534,9 @@ std::vector<std::shared_ptr<ConfigSetup>> ConfigDefinition::getServerOptions()
         std::make_shared<ConfigBoolSetup>(ConfigVal::SERVER_STORAGE_USE_TRANSACTIONS,
             "/server/storage/attribute::use-transactions", "config-server.html#storage",
             NO),
+        std::make_shared<ConfigBoolSetup>(ConfigVal::SERVER_STORAGE_SORT_KEY_ENABLED,
+            "/server/storage/attribute::enable-sort-key", "config-server.html#storage",
+            YES),
         std::make_shared<ConfigStringSetup>(ConfigVal::SERVER_STORAGE_MYSQL,
             "/server/storage/mysql", "config-server.html#storage"),
 #ifdef HAVE_MYSQL